refactor: rename package 'ports' to 'output' and 'input' for clarity
