+++ internal/domain/ports/input/token_service.go
@@ -1,4 +1,4 @@
-package ports
+package input
 
 import (
 	"context"
