+++ internal/domain/ports/output/logger.go
@@ -1,4 +1,4 @@
-package ports
+package output
 
 import "log/slog"
