+++ internal/domain/ports/output/token_repository.go
@@ -1,4 +1,4 @@
-package ports
+package output
 
 import (
 	"context"
